Aplicacion de patrones
Se aplicó el patron facade y el patrón Presentational/Container
